refactor: move migration logic into db module and update references
